Docs review: basic & builtin:
    * Remove redundant non-finite numbers section
    * Add UNit & singleton section
