+++ guide/example/example-builtin-02.kt
@@ -5,9 +5,9 @@ import kotlinx.serialization.*
 import kotlinx.serialization.json.*
 
 @Serializable
-class Data(val value: Double)                     
+class Data(val signature: Long)
 
 fun main() {
-    val data = Data(Double.NaN)
+    val data = Data(0x1CAFE2FEED0BABE0)
     println(Json.encodeToString(data))
 }
