+++ guide/example/example-builtin-03.kt
@@ -4,8 +4,13 @@ package example.exampleBuiltin03
 import kotlinx.serialization.*
 import kotlinx.serialization.json.*
 
+import kotlinx.serialization.builtins.*
+
 @Serializable
-class Data(val signature: Long)
+class Data(
+    @Serializable(with=LongAsStringSerializer::class)
+    val signature: Long
+)
 
 fun main() {
     val data = Data(0x1CAFE2FEED0BABE0)
